refactor(widgets): queue loop mode in useQueue
Move queue loop state and state setter to the api-client's useQueue hook so the state can be used by multiple components.
