+++ .changeset/young-sides-wear.md
@@ -1,0 +1,6 @@
+---
+"@sophys-web/api-client": minor
+"@sophys-web/widgets": minor
+---
+
+Move queue loop state and state setter to the api-client's useQueue hook so the state can be used by multiple components.
